Detecting number of concurrent processes

+++ lib/include/CppBuildInfo/DataParser.h
@@ -15,11 +15,11 @@ public:
     /// Gets a list with the file names (without containing the directory)
     const QStringList & getFileNames() const;
 
-    const QList<int> & getTimes() const;
+    const QList<qint64> & getTimes() const;
 
-    const QList<int> & getStartingTimes() const;
+    const QList<qint64> & getStartingTimes() const;
 
-    const QList<int> & getEndingTimes() const;
+    const QList<qint64> & getEndingTimes() const;
 
     /// Gets the number of concurrent compilation processes
     int getNComcurrentProcesses() const;
